@@ -1,5 +1,5 @@
 {
-  "name": "nodecg-react-boilerplate",
+  "name": "freshers-fair-2022-graphics",
   "version": "0.0.0",
   "description": "",
   "homepage": "",
@@ -19,11 +19,12 @@
     "compatibleRange": "^1.1.1",
     "dashboardPanels": [
       {
-        "name": "sample",
-        "title": "Sample",
-        "width": 2,
+        "name": "freshers-fair",
+        "title": "Freshers' Fair 2022",
+        "width": 4,
         "file": "dashboard.html",
-        "headerColor": "#525F78"
+        "headerColor": "#525F78",
+        "workspace": "Freshers Fair 2022"
       }
     ],
     "graphics": [
@@ -41,18 +42,26 @@
     "@babel/preset-env": "^7.15.8",
     "@babel/preset-react": "^7.13.13",
     "@babel/preset-typescript": "^7.13.0",
+    "@emotion/babel-plugin": "^11.10.2",
+    "@emotion/babel-preset-css-prop": "^11.10.0",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
     "@storybook/addon-actions": "^6.2.2",
     "@storybook/addon-essentials": "^6.2.2",
     "@storybook/addon-links": "^6.2.2",
     "@storybook/react": "^6.2.2",
+    "@typescript-eslint/eslint-plugin": "^5.38.0",
     "babel-loader": "^8.2.2",
     "cross-env": "^7.0.3",
+    "css-loader": "^6.7.1",
+    "eslint": "^8.24.0",
+    "eslint-plugin-react": "^7.31.8",
+    "file-loader": "^6.2.0",
     "fork-ts-checker-webpack-plugin": "^6.2.0",
     "html-webpack-harddisk-plugin": "^2.0.0",
     "html-webpack-plugin": "^5.3.1",
     "json-schema-to-typescript": "^10.1.4",
     "react-refresh": "^0.10.0",
+    "style-loader": "^3.3.1",
     "ts-loader": "^8.1.0",
     "typescript": "^4.2.3",
     "webpack": "^5.30.0",
@@ -63,10 +72,13 @@
     "webpack-merge": "^5.7.3"
   },
   "dependencies": {
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "@emotion/react": "^11.10.4",
+    "@types/react": "18",
+    "@types/react-dom": "18",
+    "antd": "^4.23.2",
+    "framer-motion": "^7.3.6",
+    "react": "18",
+    "react-dom": "18"
   },
   "browserslist": [
     "defaults",
